+++ client/src/components/AddPlantCard/spotlight.css
@@ -1,0 +1,5 @@
+.spotlight-card {
+  background-color: rgba(228, 198, 98, 0.637);
+  color: white;
+  padding-bottom: 20px;
+}
