+++ client/src/components/Header/Header.js
@@ -4,9 +4,9 @@ import "./header.css";
 function Header() {
   return (
     <div className="row header">
-      <div className="col-md-12 d-flex justify-content-center">
+      {/* <div className="col-md-12 d-flex justify-content-center">
         <img src='images/Trowel-logo.png' alt='trowel' />
-      </div>
+      </div> */}
     </div>
   );
 }
