+++ client/src/components/Navbar/Navbar.js
@@ -12,11 +12,15 @@ function Navbar() {
   };
   return (
     <Nav
-      style={{ backgroundColor: "rgb(206 153 4 / 65%)", height: "" }}
-      className="navbar navbar-expand-md  navbar-dark pr-lg-1 pl-sm-5 justify-content-between nav-con"
+      style={{ backgroundColor: "rgb(206 153 4 / 30%)", height: "" }}
+      className="navbar navbar-expand-md  shadow navbar-dark pr-lg-1 pl-sm-5 justify-content-between nav-con"
     >
       <Link className="navbar-brand nav-text" to="/home">
-        Trowel
+        <img
+          style={{ height: "80px" }}
+          src="images/Trowel-logo.png"
+          alt="trowel"
+        />
       </Link>
 
       <button
