--- client/src/pages/Plant/plant.css
@@ -1,5 +1,0 @@
-.spotlight-card {
-    background-color: chocolate;
-    color: white;
-    padding-bottom: 20px;
-}
